add a new portal launcher
this launcher sets the platform theme to flatpak and forwards the original
set platform theme if applicable.
when an app wants to use desktop ports it needs to use this launcher,
chances are the kf5-launch will also figure out to use flatpak though.

then it defers to kf5-launch for regular setup