add the ability to force a persistent home
by default homes are split by revision. the trouble with that is that
maaaaany applications put home locations into their configs and stuff, so
what happens is that come a new revision the config breaks because now
the home is different

as a workaround setting an env var will force the home to be set to the
persistent common dir instead. problem is that then things aren't fully
rollbackable as formats may have changed etc. however, there is not
really a good way to solve this. we can't really auto-migrate the configs
because it'd be super expensive to do so and also not error proof because
we don't know in what silly fashion a home may have been hardcoded
into the configs to begin with. and on top of that the path diff is
ultimately snapd's fault. if it bound the home so the path so it is always
static from inside the snap it'd not be a problem