do not change the PWD
we should :tm: not need the changed directory to exec properly. certainly
not on a global scale across all snaps.

changing does however break relative path resolution as now the PWD is no
longer where the command actually was run.

so, not changing the PWD forcefully like we used to is preferred.
it ought to be noted that if the snap cannot read the PWD this may
cause errors. that is in fact intentional though when the snap cannot
read some relatively